refactor: refactor to fit the new backend API
