+++ tests/orquestra/vqa/cost_function/qcbm_cost_function_test.py
@@ -4,12 +4,12 @@
 import numpy as np
 import pytest
 from orquestra.opt.gradients import finite_differences_gradient
-from orquestra.quantum.backends import SymbolicSimulator
 from orquestra.quantum.distributions import (
     MeasurementOutcomeDistribution,
     compute_clipped_negative_log_likelihood,
     compute_mmd,
 )
+from orquestra.quantum.symbolic_simulator import SymbolicSimulator
 
 from orquestra.vqa.ansatz.qcbm._qcbm import QCBMAnsatz
 from orquestra.vqa.cost_function.qcbm_cost_function import create_QCBM_cost_function
